refactor gocsv's encoders to accept custom csv.Writers
This allows us greater flexibility in writing csvs from go.  Similar to the
decoding side of this library previously, if a custom csv.Writer is
required, previously, global state would have been modified, and that's not
the best.  With this change, gocsv can accept a custom csv.Writer without
modifying global state.