Support for marshalling & unmarshalling embedded structs
The original 'Num' struct field is a now an ordered list of indexes that
can describe embedded struct fields. The last index in IndexChain is the
struct field that defines Key as a tag. Previous indexes are struct
fields that embed the struct that defines Key. IndexChain is ordered
from the outermost embedding field to the innermost field that defines
Key.