Float encoding: add tests and fix float32 support
The float32 support wasn't working because the reflection interface
always returns a float64.  To fix this, convert back to a float32 if
the type on the interface is recorded as single precision.

Note: the tests make the assumption that 'float64' and 'float32' means
ieee754 semantics precisely.  These semantics define the number of
digits that converting to decimal should produce, and rounding
behavior, such that if two implementations differ, one is right and
one is wrong.  It should therefore be safe to assume that
float64(math.Pi) becomes 3.141592653589793; no other form is allowed
by ieee754.