Default to failing on unmatched struct tags
This deviation from the upstream repository is to match the default behavior we've come to expect in our repository.

In order to not change the behavior of upstream, we elected to provide this global flag that set this behavior. This is good in that it allowed us to contribute without breaking people.

What's bad is if we try and take this as is from upstream, then we open ourselves up to forgetting to switch this flag in every package we decide to use in.

So, if we just maintain this small deviation from upstream, it allows us to stay up to date with their changes, still able to podia patches upstream, and keep our own safety intact.