Add decoding support for multiple potential CSV headers per field
This allows for decoding flexibility for a set of expected CSV headers
for struct tags. This is useful for enabling backwards-compatible CSV
  format changes.

For example, a CSV column called 'foo' can be renamed to 'Foo' without
needing a coordinated code update, as long as both 'Foo' and 'foo' are
configured in the list of expected struct tags.

Encoding always uses the first tag in the list.