MAIN-22856 - Disable base resolution in int parsing. Assume base 10.
Zero Base results in ParseInt resolving the base by examining a prefix:
https://golang.org/src/strconv/atoi.go?s=3604:3671#L60

This is not desirable for how we use gocsv, since we do not have a use
case for non-base10 CSV unmarshalling. Leaving resolution in place
breaks in cases where gocsv is used to ingest data that contains
zero-prefixed integers that are intended to be base10.