Also treat UnmarshalCSV and UnmarshalText as marshaling methods
eeb477f already prevented exported fields from nested structs from
being exposed as a fieldInfo if the nested type implements `MarshalCSV`
or `MarshalText`. I would say that this should also be the case if just
one or both of the *unmarshaling* methods are defined. Otherwise, an
exported field from a nested structure will shadow the intended
unmarshaling behavior of the top-level structure if one of its fields
has the same name as a field of the nested structure itself.
Changing this certainly doesn't break any of the existing tests.